pyproject.toml
[project]
name = "dbtool"
version = "0.1.0"
description = ""
authors = [
    {name = "Joe Sumner",email = "jonathansumner98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "psycopg2-binary (>=2.9,<3.0)",
    "rich (>=14.0,<15.0)",
    "InquirerPy (>=0.3,<0.4)",
    "humanize (>=4.0,<5.0)",
    "tomli-w (>=1.0,<2.0)",
]

[project.scripts]
dbtool = "dbtool.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"